predictor/reddit_scraper/requirements_scoring.txt
anthropic>=0.68.0
python-dotenv==1.0.0